refactor: Independent nackend
the backend in OpenVAF was previously represented by the MIR and the CFG
that were tighly coupled. However with the reworked frontend almost all
information captured by the MIR can be obtained from the HIR instead.
Furthermore type infere is now done on the HIR aswell.

As a result only the need for a CFG (and the supporting algorithems)
remain. These are implented completly seperatly from the frontend
by replacing the Input and CallType Traits/Enums with integer indexes.

Furthermore the CFG has been restructured as follows:

* Way less Spans: almost all diagnostics are emitted from the HIR now so these
  aren't needed anymore. This improves memory consumption and simplifies
  CFG transformations.

* fully statically typed: previously the exact instruction emitted for
  an RValue dependend upon the type of its operands. Now that all type
  information is known before CFG construction RValues have been
  replaced with Instruction. Each Instruction is made up of an op code,
  a list of operands and (optionally) a destination. This significantly
  simplifies implementing alogrithems for the CFG

* fully flexible: Mutable memory locations (places), function parameter
  (inputs) and callbacks are all represented by opqaue integers in the
  CFG. This allows for the flexiblity required by future work on osdic