ci: Switch to build.sr.ht
The decision to move to sourcehut was laregly driven by the great CI
workflow offered. Runners just run normal vms (not docker) and therefore
spin up much faster. After that taks are essentially just a list of
shell scripts that all share the same virtual mashine.
The CI is finally fast enough to be useful now.