ci: added caching
this is an implementation of CI caching that has 2 priorities:
* Always run vendored code in CI when secrets are available
* cache as much as possible with few rebuilds

the scheme is setup as follows:
* the dependencies are autopmatically vendored and uploaded with a git
  hook (so long as the user has access). The hash is stored in the repo
* the vendored dependencies are pull in the CI with a bash script
* aftewards cargo is invoked with the --offline command avoiding index
  fetch
* furthremore the target directory is seperetly cache
* it is uploaded to a combination of rustc version and Cargo.lock hash
* local crates are removed from Cargo.lock/target directory
  before hashing/upload to allow fairly persistant caches

these commands are setup by the newly introduced xtask binary.
xtask is intended to execute various tasks that have traditionally done
with shell scripts (because shell scripts are horrible). It is build
with cargo and can therefore only be called after the vendored caches
have been downloaded. For that purpose the binary generates
.vendor/pull.sh