refactor(HIR): unified body queries
Bodys can represent everything that has expressions and/or stmts
attached. Such as function, parameter and variable definctions and
analog bodies of modules. While the representation of these bodies is
mostly the same, the entry expr/stmts are not. In the past this was
solved by specical casing each kind of body.

Now there is only a single body struct. The entry is always a list of
stmts. For expr only bodies we generate an expr stmt for each entry
expression. During type checking expressions stmts are either treated as
an assigment with unkown dst type (aka any type) or (for parameter and
variable assigment) of a single known type. This streamlines the code a
lot (especially for validation later)