refactor: HIR validation
The inference result should be as minimal as possible and contain only
the information relevant for cfg construction (and diagnostics that
neccessarly arise when constructing this information).

As a result a lot of invariants are not checked during inference.
This PR introduces and additional pass that is run on a body after type
inferece. It checks that these invariants are upheld.

The implementation is currently incomplete as checks for
natures, disciplines and branches are missing. See ~dspom/OpenVAF#1