fix: incorrect syntax tree for function names
Due to an implementation details of how AST enum are generated, SYSFUN tokens must be
wrapped in an addtional SYS_FUN node. This change was implemented in the
AST but overlooked in the parser. This commit makes the parser create
the wrapper node. A test is also added to hir_def to ensure that bodys are
generated correctly for a practical examples as previous tests missed
this bug.