fix: variable default value body
Variables are not required to have a default value in VerilogA.
Instead a default value of 0.0/0 for real/integer variables is
automatically set. Previously the compiler would generate a missing expr
when no default is provided during body lowering. This would lead to
panics during hir_lowering (as missing exprs must only be generated in
case of invalid input). This PR fixes the body lowering for variable
default values so that an approriate const is generated instead.