refactor: separate locals and places
Previously locals could be either a temporary or a place.
This lead to complications as mapping from locals to places and back was
required.

This design was originally chosen because it made
implementing various algorithms easier by just having one global
mapping. However while optimizing these algorithms it was discovered
that temporaries and places always need to be treated separately as
temporaries can generally be considered immutable and therefore without
state while places have state and therefore require DFA.

This commit changes the definition of `Local` to always be immutable
whereas places remain their current definition.