feat: add HIR body to CFG lowering
This serves as the bridge between the backend (cfg) and the fronted
(AST/HIR). The hir_lower crate therefore is (and should always stay) the
top of the dependency tree.

During the lowering process HashSets with persistent
insert order (that can be indexed) are extremely useful to ensure that
inputs, places etc. are unique. These are provided by the IndexMap crate. However the
indices used are new type wrappers which are not support by the IndexMap
implementation. Therefore a small wrapper that adds support for new type
indices was added in the ty_index_set crate. This wrapper will likely
also be useful for other algorithms in the backend (and downstream
crates such as VerilogAE and OSDIC)