fix: remove arithmetic shift operators
This commits reverts the previous commit fix: add arithmetic shift operator.
Furthermore it completly removes the arithmetic shift operators from the
parser and syntax tree. The arithmetic shift operators are not allowed
in analog behaviour (see VAMS-LRM - 4.2.11) and therefore will not be
included in the language subset covered by OpenVAF.