feat: implement differential use algorithm
Previously derivatives were propagated by naively walking the CFG in post
order. However this approach falls short when loops are involved.
Example:

```
bar = 0;
for (i=0;i<10;i++) begin
   foo = bar + 1;
   bar = V(z);
end
res = ddx(bar,V(z));
```

Naively walking the loop backward will produce an incorrect result for
this example (res=0 instead of res=9).
When assignment to `bar` is encountered in the example, the assignment to `foo` has not
yet been reach and as such the required derivative will not be generated.

To solve this in the general case an iterative algorithm is required.
Specifically this is a data flow problem similar to live variable
analysis. However this algorithm is more sophisticated because there are
multiple tracked states for each variable (each unkown) instead of just
one (live or not live).