fix(auto_diff): unkown format for higher order derivatives
The unkown format previously implemented was ill-suited for generating
higher order derivatives. This commit introduces a slightly different
format. Each (higher order) unkown is now a pair of two unkowns:

* An unkown that represents the anti derivative of this unkown
* The last first order derivative

To arrive at an higher-order derivative of an instruction the derivative
of the instruction by the anti-derivative component is derived by the
last first-order unkown.

These unkowns essentially form an intrusive linked list.
As a result format is slightly less optimal for live derivative generation.
When the ddx operator is encountered the anti-derivative list needs to
regenerated as the same list with a different tail is required.

Because ddx operators must only appear in non-conditional behaviour
each operator is only executed at most twice. As extremely high
order derivatives (and the ddx operator itself) are fairly rare the
performance impact is likely not relevant.

However this format enables ergonomic (and correct) generation of higher order
derivatives even in edge cases and is therefore likely worth it.