feat: Implement derivative generation
This PR introduces an algorithm that can produce derivatives of
any order based upon the results of live-derivative analysis.
This algorithm is fairly complex as a lot of edge cases (such as self
assigning derivatives) have to be considered.

As a result a lot more heap allocations are required than a naive implementation
would use. To ensure efficiency allocations are aggressively reused
between multiple instructions. This sadly makes the code a bit more
difficult to grasp. Future work should look to improve this.

Furthermore benchmarks should be used to more closely monitor the
performance of this step and potentially further improve the
implementation.

This implementation is a prototype and still requires more testing.
However the current approach of manually checking control flow graph
quickly becomes too time intensive due to the amount of generated code.
Instead we should look into building full compiler integration tests here
where a derivative is actually calculated (and then compare the result
to a manual implementation)