chore: test backend
This commit introduces a comprehensive integration tests suite for the
backend. While many bugs can be covered by the various integration tests
in the various backend crates, it is often hard to test all possible
cases in these unit tests due to the complexity of the code.

The new test suite contains two kinds of tests:

* complex derivatives
  it is desierable to test that correctness of complex (especially
  higher order) derivatives. However manually checking the assembly is
  infeasible for these cases due to the sheer volume of the emitted
  code. In these integration tests the derivatives are evaluated with the
  const_eval pass and the results compared to manual implementations of
  the results.

* full compilation of real-world code
  These tests apply the full OpenVAF compilation pipline (including
  jacobian generation) to all models in the integration_tests
  direcitory. This not only helps to find obscure bugs but also diagnose
  performance problems

Note: this commit is rather large as it also contains various fixed introduced
to address the problems uncovered by the testsuite.