feat(verilogae)!: non-optional currents/voltages
In the past allowed voltages/currents to be omitted (0 was used as a
default). However this has repeatably proven problematic in practice.
See #5 for more details.

This commit changes `eval` so that an error is generated for missing
currents/voltages. This means that this is a BREAKING_CHANGE for
downstream users that relied on this behaviour.

There are cases where the old behaviour is preferable (for example
`br_tnode` for HICUM). To support those cases the `opt_voltage` and
`opt_current` attributes are added. These can be placed in front of a
branch to indicate that it's current/voltage can be omitted when calling
`eval` function. By default 0 is used however a different default value
may be supplied to the attribute. For example: `(*opt_voltage = 1*)`.