fix(verilogae): convert integer to real scalars
Addresses #4.

The `eval` function requires that all arguments are correctly typed.
However for real scalar this is counter-intuitive because both python
and Verilog-A coerce int -> float/real.

The `eval` function has been changed to perform this conversion for
real parameters. Note that float -> int conversions are generally not
lossless and therefore not performed automatically.

The conversion is only performed for scalar values and not
for numpy arrays to avoid allocating/freeing memory.