fix(sim_back): incorrect derivative metadata
Derivative unkowns are represented by opaque indices. These indices in turn map
to a `FuncRef` (a function id) that represent the appropriate call to
`ddx`. Therefore to construct an unkown we need to map any value(s) to
the appropriate call and then map that call to the unkown index.

When constructing the matrix we need to create such a mapping to find
the unkown (and therefore derivative) by a certain node voltage/branch current.
Previously the implementation only constructed a callback ID and then
went out of its way to manually convert the type to a unkown id.
This was probably done because the callback -> unkown mapping was not
available anymore during matrix construction.

This implementation happened to pass previous test cases because
all callbacks were ddx functions. All callback ids were
incidentally equal to unkown ids.

However this fails for more complex models that also use other
callbacks. This commit makes the callback -> unkown mapping available
outside of the `mir_autodiff` crate and uses that mapping to correctly
convert the callback ids to unkown ids.

This commit also adds a test case that checks which matrix entries are
generated for the HICUM model. The entries seem reasonable now.
Testers should check that the generated stamps are exactly correct.