perf(osdi): parallel compilation
Compilation of OSDI can be slow in some cases.
To address that this commit restructures OSDI so that the eval function,
the setup_model, setup_instance and initial implementation and remainder
of the code generation are all performed in parallel.

The code generation of these functions is completely independent of each
other and therefore the compilation is now only as slow as the slowest
function (usually eval).