feat(osdi): support vsrc, iprobe, idt and ddt
This commit enables OpenVAF to create internal unkowns
to support advanced Verilog-A features such as voltage sources,
current probes, switch branches and `idt`/`idtmod` and `ddt`[^1].
These unkowns are only created if needed and automatically
collapsed if they are only used sometimes.
To correctly support switch branches, the node collapsing implementation
had to be adjusted to ensure that OpenVAF can distinguish between static
node collapsing and a dynamic switch branch with a 0V voltage source.

[^1]: `ddt` already worked before however OpenVAF required that `ddt`
is only used in contribute statements with a constant factor.
Now OpenVAF automatically detects whether `ddt` is used in a linear
contribution and only creates and internal unkowns if required.