fix(osdi): Fix that $temperature always returned 0
The changes in the previous commits changed $temperature so that it was
read from memory instead of being treated as an SSA value during the eval function.

However, the wrong GEP indices ([6] instead of [0, 6]) were used.
This commit fixes that and introduces a function that automatically
generates these struct_gep indices to avoid similar bugs in the future.