feat: implement type checking $limit
Previously using $limit would cause a panic during type check.
Now OpenVAF tests that $limit is used correctly.
This is only possible because it assumes that built-in function only have
real arguments.

For now $limit(x,..) is lower the same as x in hir_lower.
However, all type checking logic is implemented so the foundations
to fully support $limit in OSDI are now available.