feat(osdi): properly implement analysis
Previously analysis always returned true because the appropriate
information was not available with OSDI.
With the latest additions to the interface the ANALYSIS_<..> flags are
set by the simulator to communicate the return value of this function.
This commit implements analysis so that is returns true for the analysis
modes listed in the language standard when the appropriate flags are
set.