chore(verilogae): enable static linking
Because Verilog-AE can be used from different programming languages, the
tool is compiled to a shared library by default that is then embedded
into the cross platform wheel by the manylinux utility.

This is not just fragile, it also blocked building on other platforms
(windows) that do not have the manylinux utility.