chore(internal): reorganize crates into subdirectories
With the addition of melange, the openvaf repo now contains 3 distinct
projects: OpenVAF, Melange and VerilogAE.
Keeping all of those crates in a single was becoming cluttered.
Instead, one subfolder has been created for each project.
The general idea is that verilogae and melange depend on openvaf but not
the other way around so they can stay in seperate directories.

Furthermore, OpenVAF uses quite a few datastructures that are not
specific to OpenVAF and could (in theory) be published to crates.io/reused
by other projects. These have been moved into the libs folder in an
attempt to further organize the code.