fix(osdi): build stdlib for every taget
Previously the OSDI buildscript only build the stdlib for the host platform
and hoped that llvm bitcode was portable enough.
However when attempting to target windows it became quickly appart that this approach does not work well in practice.
This commit changes the buildscript and inclusion mechanism to ensure that the osdi standard library always matches the target.
For this refactor to be possible the testsuite had to be refactored, because osdi got into an uncompilable state where the osdi sourcegeneration (part of osdi) could not run.
Now all source generation is part of the sourcegen crate (at the root of the repo).
Furthermore most tests have been refactored to data_tests to avoid duplicate code