refactor(openvaf): expose openvaf as a library
previously the openvaf-driver was responsible for calling the individual openvaf components (main compiler, linker).
This code was fairly simple and therefore duplicated in the melange prototype.
However this approach is error prone in the long run if the two implementations diverge.
Furthermore, we want to actually test full linking too which also requires access to openvaf as library.

In this commit the driver and the core openvaf invocations have been split into two seperate crates.
The lib crate is now reused in melange instead of the previous duplicate code.
The implementation has been refactored to be more modular so that the needs of both the driver and melange can be satified.
Finally some details that were overlooked in previous refactors were cleaned up.