fix: ensure pruned branch instructions are zapped during cfg simplfy
Branch instructions determine where to jump based upon a (bool) predicate.
During CFG simplification branches with identical destincations
`then_dst == else_dst` are replaced with jumps.
It's important that any instruction that refers to a value is
`zapped` to ensure it does not show up as a use anymore later.
A renegade use like that can cause crashes, impede opzimiations
and might hypothetically cause misscompilations (altough that has never
been observed).

During CFG simplification blocks with just a single predecessor
that has no additional sucessors are merged.
It was previously incorrectly assumed that means that the sucessor
is termined by a `Jump{ dst: sucessor }` instruction.
Therefore the removed jump was not zapped.

However a branch with identical destinations (as desribed above)
that has not yet been simplified is also possible in which case a
zap is required a conidtion to zap branches in that case is added
with this commit.