ensure there is an equation for every node
this commit fixes crashes in weird situations like when two nodes are collapsed
unconditionally. Technically, the crash here likely means that further
optimizations are possible, but I don't think these (usually buggy) models are
worth the effort (and the optimizations are likely not lossless anyway as you
wouldn't be able to probe the nodes anymore).